// types/billionaire.ts
export interface Billionaire {
  name: string
  age: number | null
  netWorth: number
  industry: string
  sourceOfWealth: string
  title: string
  organization: string
  isSelfMade: boolean
  selfMadeScore: number
  philanthropyScore: number
  residence: string
  citizenship: string
  gender: string
  maritalStatus: string
  children: string
  education: string
  year?: number | null
  wealthStatus?: string
}